requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.25.0
APScheduler>=3.10.4
pytz>=2024.1
Pillow>=10.4.0
httpx>=0.27.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
lxml>=5.2.1
facebook-sdk>=3.1.0
peewee>=3.17.3
fastapi>=0.111.0
uvicorn>=0.29.0
google-generativeai>=0.7.2
